feat: moved code to server components for improve perfomance
